requirements.txt
altair==5.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.4
boto3==1.34.159
botocore==1.34.159
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
fonttools==4.53.1
frozendict==2.4.4
gitdb==4.0.11
GitPython==3.1.43
html5lib==1.1
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==8.10.0
multitasking==0.0.11
narwhals==1.4.1
netifaces==0.11.0
numpy==2.0.1
oauthlib==3.2.0
packaging==24.1
pandas==2.2.2
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
protobuf==5.27.3
pyarrow==17.0.0
pydeck==0.9.1
pygame==2.6.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==5.4.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.0
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.37.1
systemd-python==234
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
wadllib==1.3.6
watchdog==4.0.2
webencodings==0.5.1
yfinance==0.2.41
zipp==1.0.0
